Rename plugins to prefix with `jsep`
