+++ src/index.js
@@ -1,6 +1,6 @@
 // Add default plugins:
 import jsep from './jsep.js';
-import ternary from './plugins/ternary.js';
+import ternary from './plugins/jsepTernary.js';
 
 jsep.plugins.register(ternary);
