+++ test/tests.js
@@ -2,5 +2,5 @@
 export * from './jsep.test.js';
 export * from './hooks.test.js';
 export * from './plugins.test.js';
-export * from './plugins/ternary.test.js';
-export * from './plugins/object.test.js';
+export * from './plugins/jsepTernary.test.js';
+export * from './plugins/jsepObject.test.js';
